moved nes display to separate files;
