+++ .vscode/settings.json
@@ -34,6 +34,7 @@
         "typeinfo": "cpp",
         "utility": "cpp",
         "istream": "cpp",
-        "ostream": "cpp"
+        "ostream": "cpp",
+        "retro8x16.c": "cpp"
     }
 }
